Conflicts:
	Projects/UnitTests/Classes/GameCore.cpp
	Projects/UnitTests/TemplateProjectiPhone.xcodeproj/project.pbxproj
	Projects/UnitTests/UnitTestsVS2010.vcxproj.filters
	Projects/Win32Framework/Win32Framework.vcxproj
	Sources/Internal/FileSystem/FileSystem.cpp
	Sources/Internal/FileSystem/FileSystem.h
	Sources/Internal/Platform/TemplateAndroid/DeviceInfoAndroid.cpp
	Sources/Internal/Platform/TemplateAndroid/Java/com/dava/framework/JNIDeviceInfo.java
	Sources/Internal/Platform/TemplateMacOS/DeviceInfoMacOS.mm
	Sources/Internal/Render/Highlevel/Landscape.cpp
	Sources/Internal/Render/Highlevel/VegetationRenderObject.cpp
	Sources/Internal/Render/Image/LibDdsHelper.cpp
	Sources/Internal/Render/MipmapReplacer.cpp
	Sources/Internal/Render/RenderBase.h
	Sources/Internal/Render/RenderOptions.h
	Sources/Internal/Render/Shader.h
	Sources/Internal/UI/UIControlBackground.cpp
	Sources/Internal/Utils/StringFormat.cpp
	Sources/Tools/TextureCompression/DXTConverter.cpp
	Tools/ResourceEditor/Classes/Commands2/BeastAction.cpp
	Tools/ResourceEditor/Classes/Qt/DockParticleEditor/ParticleEditorWidget.cpp
	Tools/ResourceEditor/Classes/Qt/DockProperties/PropertyEditor.h
	Tools/ResourceEditor/Classes/Qt/ImageSplitterDialog/ImageSplitterDialog.cpp
	Tools/ResourceEditor/Classes/Qt/Main/QtUtils.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/CustomColorsSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/DebugDrawSystem.h
	Tools/ResourceEditor/Classes/Qt/Scene/System/StructureSystem.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/LandscapeDialog/LandscapeDialog.cpp
	Tools/ResourceEditor/Classes/Qt/Tools/QtPropertyEditor/QtPropertyDataValidator/HeightmapValidator.cpp
	Tools/ResourceEditor/ResourceEditorQtVS2010.vcxproj
	Tools/ResourceEditor/ResourceEditorQtVS2010.vcxproj.filters

Former-commit-id: 37b9d334df8809c4894ea03b82223858d0e597b0

+++ Sources/Internal/Render/Shader.cpp
@@ -118,7 +118,10 @@ FastName attributeStrings[VERTEX_FORMAT_STREAM_MAX_COUNT] =
     FastName("inTangent"),
     FastName("inBinormal"),
     FastName("inJointWeight"),
-    FastName("inTime")
+    FastName("inTime"),
+    FastName("inPivot"),
+    FastName("inFlexibility"),
+    FastName("inAngleSinCos")
 };
 
 eShaderSemantic Shader::GetShaderSemanticByName(const FastName & name)
@@ -276,7 +279,7 @@ Shader::~Shader()
     ReleaseShaderData();
 }
 
-void Shader::ReleaseShaderData()
+void Shader::ReleaseShaderData(bool deleteShader/* = true*/)
 {
     SafeDeleteArray(attributeNames);
     //SafeDeleteArray(uniforms);
@@ -286,7 +289,8 @@ void Shader::ReleaseShaderData()
     SafeRelease(vertexShaderData);
     SafeRelease(fragmentShaderData);
     
-    DeleteShaders();
+    if (deleteShader)
+        DeleteShaders();
 
     activeAttributes = 0;
     activeUniforms = 0;
@@ -1211,7 +1215,18 @@ void Shader::BindDynamicParameters()
                 }
                 break;
             }
-
+            case PARAM_SPEED_TREE_TRUNK_OSCILLATION:
+            case PARAM_SPEED_TREE_LEAFS_OSCILLATION:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(currentUniform->shaderSemantic);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    Vector2 * param = (Vector2*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
+                    SetUniformValueByUniform(currentUniform, *param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
             case PARAM_COLOR:
             {
                 const Color & c = RenderManager::Instance()->GetColor();
@@ -1334,6 +1349,8 @@ void Shader::Lost()
     program = 0;
     activeAttributes = 0;
     activeUniforms = 0;
+    
+    ReleaseShaderData(false);
 }
 
 void Shader::Invalidate()
